Use explicit instantiation
